Save coverage data in separate files for each test suite
Save coverage data in a separate file for each test suite.  This
avoids race conditions and file corruption when running under
parallel-test.  When the merged data are required, they are sourced
from all of the coverage data files.

This is a bit rough and ready, but works for my immediate needs, and
can be tidied up later.